+++ src/Types/Procedure.hs
@@ -271,7 +271,7 @@ data FunctionQualifier c =
   CategoryFunction [c] CategoryName |
   TypeFunction [c] TypeInstanceOrParam |
   -- TODO: Does this need to allow conversion calls?
-  ValueFunction [c] (Expression c) |
+  ValueFunction [c] ValueCallType (Expression c) |
   UnqualifiedFunction
   deriving (Show)